Added a pink flash upon keypress.
TODO: Add a shockwave effect when an arrow is hit.
TODO: Generalise the input system so it will be easier to
incorporate the RaspPi GPIO pin inputs.

+++ GameObject.h
@@ -26,7 +26,7 @@ public:
     double vx;              // velocities
     double vy;
 
-    uint8_t depth;              // draw order
+    int depth;              // draw order
 
     double alpha;           // image alpha
